[operator] Allow custom labels and annotations (#1067)
* [operator] Add the ability to add `additionalLabels`

This already existed in the `opentelemetry-collector` chart, now it
is available in the `opentelemetry-operator` as well.

These labels can be useful in environments where additional labels
are recommended or required.

* [operator] Allow adding custom annotations to various objects

Similar to `additionalLabels`, but for annotations.  This is somewhat
inspired by `cert-manager`s Helm chart [1], which allows custom annotations
on essentially all Kubernetes objects.

E.g. there could be internal annotations in certain environments that
are recommended or required

[1]: https://artifacthub.io/packages/helm/cert-manager/cert-manager

* [operator] Bump chart version for custom labels and annotations support